\name{Cells}
\alias{Cells}
\alias{A0}
\alias{X0}
\alias{b0}

\title{Cell counts data from ...}

\usage{
data(Cells)
}

\description{
  A short description of the background of this set of data.
}

\value{
  \item{Cells}{Column-compressed form of the observed cell counts.  The
    first column contains the index of variables (there are 10 of them);
    the second column ..., third ... last ...}
  \item{A0}{The first iteration of the coefficient-matrix A.  A0 is in sparse
    matrix form.}
  \item{X0}{The first iteration of the initial values of X(t) used in
    the ODE system.}
  \item{b0}{The first iteration of the constant terms in the ODE system.}
}

\source{Some description of these data.}

\references{A published paper about this data.}
\keyword{datasets}  
